refactor: move the MarkdownEditor component from PoolDetails to component folder
