+++ src/templates/Manage/CreatePool/index.tsx
@@ -733,7 +733,9 @@ const CreatePool = ({ setIsCreatePool }: ICreatePoolProps) => {
           return {
             id: token.address,
             decimals: token.decimals,
-            amount: Big(poolData.tokenInAmount).mul(Number(token.allocation) / 100).toFixed(0)
+            amount: Big(poolData.tokenInAmount)
+              .mul(Number(token.allocation) / 100)
+              .toFixed(0)
           }
         })
       })
